Add script to check the status of model URLs
This commit adds a new shell script that checks the status of model URLs. The script uses curl to follow redirects and determine if each URL is live or not. If a URL is not live, it is added to an array of dead links which are printed at the end of the script.