Add support for custom environment file in link.sh script
This commit adds the ability to specify a custom environment file using the -e or --env flag when running the link.sh script. The default value is ".env", but can now be changed to any other filename. This change also updates all references to the .env file within the script to use this new variable instead of hardcoding ".env".